requirements.txt
fastapi
uvicorn
python-multipart
beautifulsoup4
requests
lxml
pydantic
python-dateutil
sqlalchemy
pymysql
mysqlclient
mysql-connector-python
pytest
pytest-asyncio
httpx
selenium
webdriver-manager
python-dotenv